Lag fixed for "model"
Modified the drawCube function for the model and had it translate the
already drawn shape to the camera position (minus cube height so we can
see) instead of redrawing the cube 60 times a second. I also edited out
the ambient lighting because it messed up your fps counter for some
reason.