Modify makefile to allow fmod(ex) linking on linux.
It does not appear that there is any standard place to put
"libfmodex.so" on linux systems. To work around this, I taught the
makefile to look in the "config.mak" file (similar to the git project)
for extra variable definitions, and the contents of MY_LIBS is appended
to the LIBS variable. It may be desirable to move other nonstandard
location-type information to this file.

For reference, the so file on my system is in /opt/fmodex/api/lib, so my
config.mak file has a line:
MY_LIBS=-L/opt/fmodex/api/lib