Add an example file for config.mak
This example file should provide some insight as to the correct use of
the config.mak file. It is designed to provide a starting point for
a machine-specific config.mak file, so it need not be written from
scratch by anyone needing to tweak the program to build on their system.

It should be updated whenever any variables that are machine-specific
are added to the makefile.